requirements/base.txt
Django==4.1.4
django-admin-rangefilter==0.9.0
django-ckeditor==6.5.1
django-environ==0.9.0
django-mptt==0.14.0
Pillow==9.3.0
psycopg2-binary==2.9.5
six==1.16.0
